+++ Transport_Compare.cpp
@@ -1,5 +1,5 @@
 #include "Transport.h"
-int WPRatio(Transport* lg);
+float WPRatio(Transport* lg);
 bool Compare(Transport* first, Transport* second)
 {
 	return WPRatio(first) < WPRatio(second);
